+++ newclient/src/modules/address/handlers/AddressPendingTx/AddressPendingTx.vue
@@ -1,7 +1,7 @@
 <template>
     <v-card color="white" flat class="pb-2">
         <app-table-title :has-pagination="showPagination" :title="getTitle" :page-type="pageType" page-link="">
-            <template v-slot:pagination v-if="showPagination">
+            <template v-if="showPagination" #pagination>
                 <app-paginate
                     :class="$vuetify.breakpoint.smAndDown ? 'pt-3' : ''"
                     :total="totalPages"