+++ newclient/src/modules/favorites/models/ErrorMessagesFav.ts
@@ -1,0 +1,6 @@
+export enum ErrorMessagesFav {
+    addressCheck = 'message.error.favorites.address-check',
+    ethBalance = 'message.error.favorites.eth-balance',
+    contract = 'message.error.favorites.contract',
+    rewards = 'message.error.favorites.rewards'
+}