+++ newclient/src/modules/favorites/models/FavApolloRefetch.ts
@@ -1,0 +1,7 @@
+import { DocumentNode } from 'graphql'
+import { checkAddressVariables } from '@app/modules/favorites/handlers/apolloTypes/checkAddress'
+
+export interface CheckAddressRefetch {
+    query: DocumentNode
+    variables: checkAddressVariables
+}